fix: migration and improve ux
